refactor(desktop): one home for per-room session state; drop dead indirection
RoomTranscriptService had quietly absorbed four jobs: transcript reads,
attention mode, monitor health, and view-instance ownership. The last
three are one responsibility — this conversation's relationship to a
room, not the room's content — and now live together in RoomSession
(attention.py), which the service composes and the server addresses
directly. The untracked-room fallback follows the system default
(user_first), so a widget remounted after a process restart is not
greeted by an armed staleness alarm for a loop nobody promised.

Also deletes the workflow-operation registry: an enum, a type alias, a
dict, and a dispatch loop that served exactly one operation consumed by
two callers, which now simply call the function.
